+++ jarvis/_preexec/keywords_handler.py
@@ -3,6 +3,7 @@ import warnings
 from collections import OrderedDict
 from typing import NoReturn
 
+import inflect
 import yaml
 
 from jarvis.modules.builtin_overrides import ordered_dump, ordered_load
@@ -17,6 +18,8 @@ if not os.path.isdir(models.fileio.root):
 def rewrite_keywords() -> NoReturn:
     """Loads keywords.yaml file if available, else loads the base keywords module as an object."""
     keywords_src = OrderedDict(**keywords.keyword_mapping(), **conversation.conversation_mapping())
+    # WATCH OUT: for changes in keyword/function name
+    keywords_src['events'] = [models.env.event_app.lower(), inflect.engine().plural(models.env.event_app)]
     if os.path.isfile(models.fileio.keywords):
         with open(models.fileio.keywords) as dst_file:
             try: